+++ config/templates/GetnetChileConfig.php
@@ -7,10 +7,15 @@ use PlacetoPay\PaymentMethod\Constants\Environment;
 abstract class CountryConfig
 {
     public const CLIENT_ID = 'getnet-chile';
+
     public const CLIENT_URI = 'https://www.getnet.cl';
+
     public const CLIENT = 'Getnet';
+
     public const IMAGE = 'https://banco.santander.cl/uploads/000/029/870/0620f532-9fc9-4248-b99e-78bae9f13e1d/original/Logo_WebCheckout_Getnet.svg';
+
     public const COUNTRY_CODE = 'CL';
+
     public const COUNTRY_NAME = 'Chile';
 
     public static function getEndpoints(): array